Move helpers from Plugins to PluginInfoService
